src/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
psycopg2-binary==2.7.7
SQLAlchemy==1.3.0
APScheduler==3.6.
requests==2.22.0
gunicorn==20.0.4
